# In-memory cache for NIR linking