This commit enables scala-native to cache the NIR link results
based on input: the hash of

- `config.artifactName`
- `entires: Seq[Global]`
- timestamps of the classPath

Since we use `config.artifactName`, we have the same problem as
https://github.com/scala-native/scala-native/pull/3283#issuecomment-1546274083

Related https://github.com/scala-native/scala-native/issues/3282

Before

```
[info] Linking (multithreading false) (596 ms)
[info] Checking intermediate code (74 ms)
[info] Dumping intermediate code (linked) (230 ms)
[info] Discovered 778 classes and 4879 methods
[info] Optimizing (debug mode) (546 ms)
[info] Checking intermediate code (33 ms)
[info] Dumping intermediate code (optimized) (125 ms)
[info] Dumping intermediate code (lowered) (173 ms)
[info] Generating intermediate code (512 ms)
[info] Produced 39 files
[info] Compiling to native code (209 ms)
[info] Linking native code (immix gc, none lto) (3 ms)
[info] Total (2370 ms)
[success] Total time: 4 s, completed May 15, 2023, 9:43:49 PM
sbt:scala-native> sandbox2_13/nativeLink
[info] Linking (multithreading false) (323 ms)
[info] Checking intermediate code (49 ms)
[info] Dumping intermediate code (linked) (120 ms)
[info] Discovered 778 classes and 4879 methods
[info] Optimizing (debug mode) (357 ms)
[info] Checking intermediate code (15 ms)
[info] Dumping intermediate code (optimized) (96 ms)
[info] Dumping intermediate code (lowered) (182 ms)
[info] Generating intermediate code (363 ms)
[info] Produced 39 files
[info] Compiling to native code (166 ms)
[info] Linking native code (immix gc, none lto) (4 ms)
[info] Total (1507 ms)
[success] Total time: 2 s, completed May 15, 2023, 9:43:54 PM
```

After

```
sbt:scala-native> sandbox2_13/nativeLink
[info] Linking (multithreading false) (717 ms)
[info] Checking intermediate code (105 ms)
[info] Dumping intermediate code (linked) (223 ms)
[info] Discovered 778 classes and 4879 methods
[info] Optimizing (debug mode) (549 ms)
[info] Checking intermediate code (36 ms)
[info] Dumping intermediate code (optimized) (129 ms)
[info] Dumping intermediate code (lowered) (207 ms)
[info] Generating intermediate code (510 ms)
[info] Produced 39 files
[info] Compiling to native code (172 ms)
[info] Linking native code (immix gc, none lto) (3 ms)
[info] Total (2486 ms)
[success] Total time: 4 s, completed May 15, 2023, 9:41:43 PM
sbt:scala-native> sandbox2_13/nativeLink
[info] NIR linking skipped due to cache hit.
[info] Linking (multithreading false) (12 ms)
[info] Checking intermediate code (22 ms)
[info] Dumping intermediate code (linked) (134 ms)
[info] Discovered 778 classes and 4879 methods
[info] Optimizing (debug mode) (275 ms)
[info] Checking intermediate code (15 ms)
[info] Dumping intermediate code (optimized) (110 ms)
[info] Dumping intermediate code (lowered) (160 ms)
[info] Generating intermediate code (301 ms)
[info] Produced 39 files
[info] Compiling to native code (165 ms)
[info] Linking native code (immix gc, none lto) (1 ms)
[info] Total (1045 ms)
```