@tanishiking 2023-05-22
Closing for https://github.com/scala-native/scala-native/pull/3287